reorganize interpreter
