+++ .pre-commit-config.yaml
@@ -12,29 +12,29 @@ repos:
         additional_dependencies: [black==19.3b0]
         language_version: python3.7
 -   repo: https://github.com/asottile/seed-isort-config
-    rev: v1.7.0
+    rev: v1.9.0
     hooks:
     -   id: seed-isort-config
         args: [--application-directories, "src:."]
 -   repo: https://github.com/pre-commit/mirrors-isort
-    rev: v4.3.15
+    rev: v4.3.20
     hooks:
     -   id: isort
 -   repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v2.1.0
+    rev: v2.2.3
     hooks:
     -   id: trailing-whitespace
     -   id: end-of-file-fixer
     -   id: check-yaml
     -   id: debug-statements
     -   id: flake8
-        additional_dependencies: ["flake8-bugbear == 18.8.0"]
+        additional_dependencies: ["flake8-bugbear == 19.3.0"]
         language_version: python3.7
 -   repo: https://github.com/asottile/pyupgrade
-    rev: v1.12.0
+    rev: v1.17.1
     hooks:
     -   id: pyupgrade
 -   repo: https://github.com/pre-commit/pygrep-hooks
-    rev: v1.2.0
+    rev: v1.4.0
     hooks:
     -   id: rst-backticks
