+++ tox.ini
@@ -95,7 +95,7 @@ commands = echo {posargs}
 [flake8]
 max-complexity = 22
 max-line-length = 99
-ignore = E203, W503, C901, E402
+ignore = E203, W503, C901, E402, B011
 
 [pep8]
 max-line-length = 99
